frontend/requirements.txt
streamlit>=1.30.0
pydantic>=2.0.0
ruamel.yaml>=0.18.0
requests>=2.31.0
